requirements.txt
flask
gunicorn
tensorflow==2.20.0
numpy
pillow
opencv-python